requirements.txt
Flask
pycryptodome
# Opcional: dependendo de como você implementou os serviços XML
# xmltodict
# lxml